Alter system include path for packaged python2.7 handlers.
When deploying a function that specifies `_serverless_handler.py` in a package
subdirectory, the nested directory is not included in the enviroment's
import path. Consequently, calling `from handler import handler` fails.
In order to import from the nested directory, the directory needs to be
explicitly appended with `sys.path`. Update `_afterCopyDir` to include
the `sys` module and append the directory to the import path before
importing local modules.